fix: move tag command to subcommand of nodes
